new DOCK pipeline version! strainfilter optimization

+++ build-3d.bash
@@ -13,40 +13,27 @@
 # opt: WORK_DIR
 
 cwd=$PWD
-WORK_DIR=${WORK_DIR-/dev/shm}
+export WORK_DIR=${WORK_DIR-/dev/shm}
 
-if ! [ -d $WORK_DIR/strainfilter_noH ]; then
-        echo "copying strainfilter files"
-        cp -r $HOME/soft/strainfilter_noH $WORK_DIR
+if ! [ -f $WORK_DIR/DOCK.1/.done ]; then
+	# use flock to copy software over in a thread-safe way
+	flock -w 0 /tmp/dock_build.lock $cwd/cp_dock_build.bash
 fi
-
-if ! [ -d $WORK_DIR/DOCK ]; then
-        echo "copying DOCK files"
-        time cp $HOME/soft/DOCK.tar.gz $WORK_DIR/DOCK.tar.gz
-        echo "untarring DOCK files"
-        cd $WORK_DIR
-        time tar -xzf DOCK.tar.gz
-fi
+while ! [ -f $WORK_DIR/DOCK.1/.done ]; do
+	echo "waiting for DOCK..."
+	sleep 2
+done
 
 if ! [ -f /tmp/lig_build_py3-3.7/.done ]; then
-        function build_py_env {
-        mkdir /tmp/lig_build_py3-3.7
-        echo "copying python env"
-        time cp $HOME/soft/lig_build_py3-3.7.tar.gz /tmp/lig_build_py3-3.7
-        echo "untarring python env"
-        cd /tmp/lig_build_py3-3.7
-        time tar -xzf lig_build_py3-3.7.tar.gz
-        echo "done!" > .done
-        }
+	# only one thread at a time should be performing a copy/untar operation
         flock -w 0 /tmp/lig_build.lock $cwd/cp_lig_build.bash
 fi
-
 while ! [ -f /tmp/lig_build_py3-3.7/.done ]; do
         echo "waiting for python environment..."
         sleep 2
 done
 
-export DOCKBASE=$WORK_DIR/DOCK
+export DOCKBASE=$WORK_DIR/DOCK.1
 
 function log {
     echo "[build-3d $(date +%X)]: $@"
@@ -61,7 +48,7 @@ function mkcd {
 
 if [ -f $OUTPUT/$SLURM_ARRAY_TASK_ID.tar.gz ]; then
     log "results already present in $OUTPUT_BASE for this job, exiting..."
-    mv /scratch/batch_3d*$SLURM_ARRAY_JOB_ID*$SLURM_ARRAY_TASK_ID* $LOGGING
+    mv /tmp/slurm*$SLURM_ARRAY_JOB_ID*$SLURM_ARRAY_TASK_ID* $LOGGING
     exit
 fi
 
@@ -82,16 +69,19 @@ log `wc -l $SLURM_ARRAY_TASK_ID`
 
 # this will contain all the necessary exports, software, etc.
 source $cwd/env_new_lig_build.sh
-export DOCKBASE=$WORK_DIR/DOCK
 export DEBUG=TRUE
+
 ${DOCKBASE}/common/on-one-core-py3 - ${DOCKBASE}/ligand/generate/build_database_ligand_strain_noH.sh -H 7.4 --no-db ${SLURM_ARRAY_TASK_ID}
 
 log "finished build job on $WORK_BASE/$SLURM_ARRAY_TASK_ID"
 
+echo "1" > finished/.dock_version
 tar -czf $SLURM_ARRAY_TASK_ID.tar.gz finished/*
 mv $SLURM_ARRAY_TASK_ID.tar.gz $OUTPUT
 mv /tmp/slurm*$SLURM_ARRAY_JOB_ID*$SLURM_ARRAY_TASK_ID* $LOGGING
 
 cd $cwd
 
+if [ -z $SKIP_DELETE ]; then
 rm -r $WORK_BASE
+fi
